Addressed several issues with IR user experience
- Selecting hue in color mode no longer wraps around
  - This should make it easy for users to select pure red
- Color editor now glows instead of flashes when active
  - This should prevent epileptic responses while editing colors(sorry D:)
- Number pad changes value type while in palette mode
  - Makes it way less confusing to change palette configuration
- While in color mode, the number pad changes the color to a preset.
  - The preset color can be changed with the color editor

+++ src/ledcontroller.cpp
@@ -140,9 +140,11 @@ void LEDController::setGradient(const CRGB& start, const CRGB& end) {
 }
 
 uint8_t LEDController::calculateGradientPixel(const int i, const uint8_t final, const uint8_t initial) {
-    int16_t change = final - initial;
-    int changePerPixel = change / (LED_COUNT - 1);
-    int result = initial + (changePerPixel * i);
+    // final: 127
+    // initial: 255
+    float change = final - initial;
+    float changePerPixel = change / (LED_COUNT - 1);
+    int16_t result = initial + (changePerPixel * i);
     if(result < 0) {
         result *= -1;
     }
